fixed cyclical import error, pr fixes
